+++ packages/patrol/darwin/Classes/AutomatorServer/Automator/MacOSAutomator.swift
@@ -73,6 +73,12 @@
       }
     }
 
+    func tapAt(coordinate vector: CGVector, inApp bundleId: String) throws {
+      try runAction("tapAt") {
+        throw PatrolError.methodNotImplemented("tapAt")
+      }
+    }
+
     func enterText(
       _ data: String,
       byText text: String,