+++ packages/patrol/lib/src/native/native_automator.dart
@@ -557,6 +557,28 @@ class NativeAutomator {
     );
   }
 
+  /// Taps at a given [location].
+  ///
+  /// [location] must be in the inclusive 0-1 range.
+  Future<void> tapAt(Offset location, {String? appId}) async {
+    assert(location.dx >= 0 && location.dx <= 1);
+    assert(location.dy >= 0 && location.dy <= 1);
+
+    // Needed for an edge case observed on Android where if a newly opened app
+    // updates its layout right after being launched, tapping without delay fails
+    await Future<void>.delayed(const Duration(milliseconds: 5));
+
+    await _wrapRequest('tapAt', () async {
+      await _client.tapAt(
+        TapAtRequest(
+          x: location.dx,
+          y: location.dy,
+          appId: appId ?? resolvedAppId,
+        ),
+      );
+    });
+  }
+
   /// Enters text to the native view specified by [selector].
   ///
   /// If the text field isn't immediately visible, this method waits for the